Add files via upload
This saves models with two options: one, you make a model architecture, and can save it for upload to a configuration file. The other one takes a model and adds input and output layers to match the formatting of the input and output dimensions. You still need to add weights, optimizers, etc. Need to make it use dby configuration setup.